Amended missing assetId renaming
