+++ src/app/routes/cpsAsset/utils/getAssetUri.js
